requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
mediapipe>=0.10.0
fastapi>=0.95.0
uvicorn>=0.22.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tensorboard>=2.13.0
hydra-core>=1.3.0
tqdm>=4.65.0
Pillow>=9.5.0
pytest>=7.3.0
streamlit>=1.22.0
python-multipart>=0.0.6
